+++ skywalking/__init__.py
@@ -37,6 +37,7 @@ class Component(Enum):
     RabbitmqProducer = 52
     RabbitmqConsumer = 53
     Elasticsearch = 47
+    HBase = 94
     Urllib3 = 7006
     Sanic = 7007
     AioHttp = 7008